changed data types for budget, boxOffice, and format for dates. Updated film list view model and the view. Updated DB.

+++ FilmWebProject/Core/Models/Film.cs
@@ -12,10 +12,10 @@ namespace FilmWebProject.Core.Models
         public string LongDescription { get; set; }
         public string Production { get; set; }
         public DateTime Release { get; set; }
-        public double Budget { get; set; }
+        public decimal Budget { get; set; }
         public double Score { get; set; }
         public string Studio { get; set; }
-        public double BoxOffice { get; set; }
+        public decimal BoxOffice { get; set; }
 
         public ICollection<Genre> Genres { get; set; }
         public ICollection<Review> Reviews { get; set; }
